# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "News API",
        "description": "API для управління новинами з блоковою структурою контенту.",
        "version": "1.0.0"
    },
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Реєстрація нового користувача",
                "operationId": "8e4e3cb7b46126f4696379d3c7eeb4ad",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Макс"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "max@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    },
                                    "avatar": {
                                        "description": "Аватар (файл)",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "bio": {
                                        "type": "string",
                                        "example": "Про мене"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Успішна реєстрація"
                    },
                    "422": {
                        "description": "Помилка валідації"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Авторизація користувача",
                "operationId": "222b2625838e3c57e017987f73598fda",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "max@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успішний вхід"
                    },
                    "401": {
                        "description": "Невірні облікові дані"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Вихід з акаунта",
                "operationId": "ad65cbbd4e9f201619eae184a5961a98",
                "responses": {
                    "200": {
                        "description": "Успішний вихід"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/new_blocks": {
            "post": {
                "tags": [
                    "News Blocks"
                ],
                "summary": "Створити блок контенту для новини",
                "operationId": "1f8da008a9daf20de31c96ad86afba5e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "news_id",
                                    "content[type]"
                                ],
                                "properties": {
                                    "news_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "content[type]": {
                                        "type": "string",
                                        "enum": [
                                            "text",
                                            "image",
                                            "text_image_right",
                                            "text_image_left"
                                        ],
                                        "example": "text"
                                    },
                                    "content[text]": {
                                        "type": "string",
                                        "example": "Текст блоку"
                                    },
                                    "content[image]": {
                                        "description": "Картинка блоку",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "position": {
                                        "type": "integer",
                                        "example": 0
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Блок створено"
                    },
                    "401": {
                        "description": "Неавторизований"
                    },
                    "403": {
                        "description": "Заборонено"
                    },
                    "422": {
                        "description": "Помилка валідації"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/new_blocks/{id}": {
            "post": {
                "tags": [
                    "News Blocks"
                ],
                "summary": "Оновити блок контенту (POST з _method=PUT для файлів)",
                "operationId": "b37e5416a545dbee042ff1c12f368f42",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID блоку",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "_method": {
                                        "description": "Спуфінг методу",
                                        "type": "string",
                                        "example": "PUT"
                                    },
                                    "content[type]": {
                                        "type": "string",
                                        "enum": [
                                            "text",
                                            "image",
                                            "text_image_right",
                                            "text_image_left"
                                        ]
                                    },
                                    "content[text]": {
                                        "type": "string"
                                    },
                                    "content[image]": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "position": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Блок оновлено"
                    },
                    "401": {
                        "description": "Неавторизований"
                    },
                    "403": {
                        "description": "Заборонено"
                    },
                    "422": {
                        "description": "Помилка валідації"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "News Blocks"
                ],
                "summary": "Видалити блок контенту",
                "operationId": "e99f074fe414890280840f22fcc8406e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID блоку",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Блок видалено"
                    },
                    "401": {
                        "description": "Неавторизований"
                    },
                    "403": {
                        "description": "Заборонено"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Введіть токен, отриманий при логіні",
                "bearerFormat": "sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}